small refactoring

+++ MusicDownloader/src/org/kreed/musicdownloader/interfaces/LoadPercentageInterface.java
@@ -3,7 +3,7 @@ package org.kreed.musicdownloader.interfaces;
 
 public interface LoadPercentageInterface {
 	
-	public void insertProgress (double progress, long downloadId );
+	public void insertProgress (long progress, long downloadId );
 		
 	public void setFileUri(String uri, long downloadId);
